refactor: SelectItem 위치 SelecItemArea/로 이동
