+++ frontend/src/components/Header/SearchBar/SelectItemArea/SelectItem/SelecItem.style.ts
